+++ package.json
@@ -14,6 +14,7 @@
     "@babel/core": "7.8.4",
     "@babel/preset-env": "7.8.4",
     "@babel/preset-react": "7.8.3",
+    "@types/react": "16.9.32",
     "axios-mock-adapter": "1.17.0",
     "babel-jest": "25.1.0",
     "babel-loader": "8.0.6",
@@ -26,6 +27,9 @@
     "prop-types": "15.7.2",
     "react-test-renderer": "16.12.0",
     "redux-mock-store": "1.5.4",
+    "ts-loader": "6.2.2",
+    "typescript": "3.8.3",
+    "utility-types": "3.10.0",
     "webpack": "4.41.5",
     "webpack-cli": "3.3.10",
     "webpack-dev-server": "3.10.2"
