+++ src/components/app/app.tsx
@@ -1,21 +1,21 @@
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 import {connect} from "react-redux";
 import {Router, Route, Switch} from "react-router-dom";
-import Main from "../main/main.jsx";
-import MoviePage from "../movie-page/movie-page.jsx";
-import SignIn from "../sign-in/sign-in.jsx";
-import AddReview from "../add-review/add-review.jsx";
-import MyList from "../my-list/my-list.jsx";
+import Main from "../main/main";
+import MoviePage from "../movie-page/movie-page";
+import SignIn from "../sign-in/sign-in";
+import AddReview from "../add-review/add-review";
+import MyList from "../my-list/my-list";
 import withIsValid from '../../hocs/with-is-valid/with-is-valid';
 import withActiveItem from '../../hocs/with-active-item/with-active-item';
-import {getFilteredMovieCards, isAppLoading, getPromoFilm} from "../../reducer/films/selectors.js";
-import Loader from "../loader/loader.jsx";
+import {getFilteredMovieCards, isAppLoading, getPromoFilm} from "../../reducer/films/selectors";
+import Loader from "../loader/loader";
 import {Operation as UserOperation} from "../../reducer/user/user";
 import {Operation as CommentsOperation} from "../../reducer/comments/comments";
-import history from "../../history.js";
-import PrivateRoute from "../private-route/private-route.jsx";
-import FullVideoPlayer from "../full-video-player/full-video-player.jsx";
+import history from "../../history";
+import PrivateRoute from "../private-route/private-route";
+import FullVideoPlayer from "../full-video-player/full-video-player";
 import withFullPlayer from "../../hocs/with-full-player/with-full-player";
 import {getMovieById, getSimilarFilms} from "../../utils";
 import {isAuth} from "../../reducer/user/selectors";
