+++ src/components/genre-list/genre-list.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import GenreList from "./genre-list.jsx";
+import GenreList from "./genre-list.js";
 
 const ALL_GENRES = `All genres`;
