+++ src/components/genre-list/genre-list.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 
 const GenreList = (props) => {
