+++ src/components/loader/loader.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import Loader from "./loader.jsx";
+import Loader from "./loader.js";
 
 it(`Loader component should render correct`, () => {
   const tree = renderer
