+++ src/components/main/main.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import {Main} from "./main.jsx";
+import {Main} from "./main.js";
 
 const ALL_GENRES = `All genres`;
