+++ src/components/movie-page-details/movie-page-details.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 import {formatTime} from "../../utils";
