+++ src/components/movie-page-overview/movie-page-overview.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 import {getRatingLevel} from "../../utils";
