+++ src/components/movie-page-reviews/movie-page-reviews.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 import {connect} from 'react-redux';
 import moment from 'moment';
