+++ src/components/show-more/show-more.e2e.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, {shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import ShowMore from "./show-more.jsx";
+import ShowMore from "./show-more.js";
 
 Enzyme.configure({
   adapter: new Adapter(),
