+++ src/components/sign-in/sign-in.e2e.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Enzyme, {mount} from 'enzyme';
-import SignIn from "./sign-in.jsx";
+import SignIn from "./sign-in.js";
 import Adapter from "enzyme-adapter-react-16/build";
 
 jest.mock(`react-router-dom`, () => ({Link: `a`}));
