+++ src/components/sign-in/sign-in.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from "react";
 import PropTypes from 'prop-types';
 import {Link} from 'react-router-dom';
