+++ src/components/small-movie-card/small-movie-card.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import SmallMovieCard from "./small-movie-card.jsx";
+import SmallMovieCard from "./small-movie-card.js";
 
 const film = {
   id: 0,
