+++ src/components/small-movie-card/small-movie-card.tsx
@@ -1,6 +1,6 @@
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
-import VideoPlayer from "../video-player/video-player.jsx";
+import VideoPlayer from "../video-player/video-player";
 import {Link} from "react-router-dom";
 
 const SmallMovieCard = (props) => {
