+++ src/components/tabs/tabs.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 
 const tabs = [`Overview`, `Details`, `Reviews`];
