+++ src/components/user-block/user-block.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import * as React from "react";
 import PropTypes from "prop-types";
-import {getAuthUser, isAuth} from "../../reducer/user/selectors.js";
+import {getAuthUser, isAuth} from "../../reducer/user/selectors";
 import {connect} from "react-redux";
 import {Link} from 'react-router-dom';
