+++ src/hocs/with-full-player/with-full-player.tsx
@@ -1,4 +1,4 @@
-import React, {createRef} from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 import {formatPlayerTime} from "../../utils";
 
@@ -7,7 +7,7 @@ const withFullPlayer = (Component) => {
     constructor(props) {
       super(props);
 
-      this._videoRef = createRef();
+      this._videoRef = React.createRef();
 
       this.state = {
         isPlaying: false,
