+++ src/reducer/comments/selectors.js
@@ -1,4 +1,4 @@
-import NameSpace from "../name-space.js";
+import NameSpace from "../name-space";
 
 export const getComments = (state) => state[NameSpace.COMMENTS].comments;
